use util method to validate request body
